# Conflicts:
#	src/ccmain/pageiterator.cpp
#	src/ccmain/paragraphs.cpp
#	src/ccutil/genericvector.h
#	src/ccutil/serialis.h
#	src/lstm/series.cpp
#	src/training/merge_unicharsets.cpp

+++ src/ccutil/genericvector.h
@@ -75,7 +75,7 @@ public:
   }
 
   // Return the size used.
-  int size() const {
+  unsigned size() const {
     return size_used_;
   }
   // Workaround to avoid g++ -Wsign-compare warnings.
@@ -317,7 +317,7 @@ inline bool SaveDataToFile(const GenericVector<char> &data, const char *filename
   if (fp == nullptr) {
     return false;
   }
-  bool result = static_cast<int>(fwrite(&data[0], 1, data.size(), fp)) == data.size();
+  bool result = fwrite(&data[0], 1, data.size(), fp) == data.size();
   fclose(fp);
   return result;
 }
@@ -382,7 +382,7 @@ public:
   }
   PointerVector<T> &operator+=(const PointerVector &other) {
     this->reserve(this->size_used_ + other.size_used_);
-    for (int i = 0; i < other.size(); ++i) {
+    for (unsigned i = 0; i < other.size(); ++i) {
 #if defined(_DEBUG) && defined(_CRTDBG_REPORT_FLAG)
       this->push_back(new (_CLIENT_BLOCK, __FILE__, __LINE__) T(*other.data_[i]));
 #else
@@ -706,7 +706,7 @@ void GenericVector<T>::operator+=(const T &t) {
 template <typename T>
 GenericVector<T> &GenericVector<T>::operator+=(const GenericVector &other) {
   this->reserve(size_used_ + other.size_used_);
-  for (int i = 0; i < other.size(); ++i) {
+  for (unsigned i = 0; i < other.size(); ++i) {
     this->operator+=(other.data_[i]);
   }
   return *this;
@@ -782,7 +782,7 @@ bool GenericVector<T>::read(TFile *f, std::function<bool(TFile *, T *)> cb) {
       }
     }
   } else {
-    if (f->FReadEndian(data_, sizeof(T), size_used_) != size_used_) {
+    if (f->FReadEndian(data_, sizeof(T), size_used_) != static_cast<unsigned>(size_used_)) {
       return false;
     }
   }
